feat(api-class): generalize use of int64 to prevent castings everywhere

+++ class-app/internal/api-class/datamodel/model.go
@@ -8,8 +8,8 @@ import (
 type Class struct {
 	bun.BaseModel `bun:"table:class,alias:class"`
 
-	ID      int       `bun:"id,pk,autoincrement"`
-	Number  int       `bun:",notnull,type:int"`
+	ID      int64     `bun:"id,pk,autoincrement"`
+	Number  int64     `bun:",notnull,type:int"`
 	Year    string    `bun:",notnull,type:varchar(256)"`
 	Created time.Time `bun:",notnull,default:current_timestamp"`
 	Updated time.Time `bun:",notnull,default:current_timestamp"`
